requirements.txt
fastapi==0.110.0
python-dotenv==1.0.1
gunicorn==21.2.0
uvicorn==0.27.1

mongo-migrate==0.1.2
motor==3.6.0

# Authentication dependencies
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic[email]==2.5.0
